Fix: subprocess pipe leak.
Our use of the childprocess gem was causing a pipe leak, when running
disc presence tests -- though not for any other commands. I didn't have
time to debug (gem source looked OK), so back to native Process.spawn.
This invocation may not work on Windows, but something for the future.